+++ lib/src/constraints/file_extenstion.dart
@@ -1,7 +1,7 @@
 class FileExtension {
   static const String feature = '.feature';
-  static const String generatedTest = '.bdd.test.dart';
-  static const String generatedScenarios = '.bdd.scenarios.dart';
-  static const String renamedTest = '.bdd_test.dart';
-  static const String renamedScenarios = '.bdd_scenarios.dart';
+  static const String generatedTest = '.bdd_test.dart';
+  static const String generatedScenarios = '.bdd_scenarios.dart';
+  // static const String renamedTest = '.bdd_test.dart';
+  // static const String renamedScenarios = '.bdd_scenarios.dart';
 }