+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - add custom CLI
 - `dart run bdd remane` to remove `.g` from generated files extension
 - `dart run bdd build` to generate files from feature files
+- [] separete command option and config options
 
 ## 0.2.0